Subindo a camada de servico da entidade frete

+++ entregas/src/main/java/br/com/gestao/entregas/entities/veiculo/DadosListagemVeiculo.java
@@ -1,7 +1,7 @@
 package br.com.gestao.entregas.entities.veiculo;
 
 
-public record DadosListagemVeiculo(Long proprietario, String marca, Tipo tipo, Double peso, String placa) {
+public record DadosListagemVeiculo(Long proprietario, String marca, Tipo tipo, int peso, String placa) {
 
     public DadosListagemVeiculo(Veiculo veiculo){
         this(veiculo.getProprietario().getId(), veiculo.getMarca(), veiculo.getTipo(), veiculo.getPeso(), veiculo.getPlaca());
